add: attendence of students, and teacher profile

+++ src/main/java/com/gaipov/talim_crm/dto/PaymentDto.java
@@ -14,7 +14,7 @@ public class PaymentDto {
     private Long id;
     private Long student;
     private Long groupId;
-    private BigDecimal sum;
+    private Double sum;
     private PaymentType paymentType;
     private PaymentStatus paymentStatus;
     private LocalDate created_at;
